requirements.txt
fastapi
uvicorn
firebase-admin
google-generativeai
langchain
langchain-google-genai
transformers
torch
uagents
python-dotenv
pydantic
fetch-ai-client